confusionmatrix-genai-main/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
google-genai>=1.0.0
gTTS==2.4.0
requests==2.31.0
Pillow>=10.2.0
Werkzeug==3.0.1
markdown==3.5.1
Pygments==2.17.2
numpy>=1.24.0
faiss-cpu>=1.7.4
